PostFeed is now it's own component
